Consolidate filters reproductions
